// type.ts
export interface Attraction {
  id: string;
  name: string;
  image: string;
  address: string;
  latitude: number;
  longitude: number;
}

export interface Coffee {
  id: string;
  name: string;
  image: string;
  address: string;
  phone: string;
  latitude: number;
  longitude: number;
}

export interface Festival {
  id: string;
  name: string;
  month: string;
  image: string;
}

export interface Restaurant {
  id: string;
  name: string;
  image: string;
  address: string;
  phone: string;
  latitude: number;
  longitude: number;
}

export interface Temple {
  id: string;
  name: string;
  image: string;
  address: string;
  latitude: number;
  longitude: number;
}